update: series scatter refactor
